+++ APJExtensions/UIColorExtensionTest.swift
@@ -18,7 +18,7 @@ class UIColorExtensionTest: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
 
-    func testColorFromHex() {
+    func testColorFromHexString() {
         let whiteCGCloroComponents: Array<CGFloat> = [1.0, 1.0, 1.0, 1.0],
             whiteColor = UIColor.colorFrom(hex: "#FFFFFF")
         XCTAssertEqual(whiteCGCloroComponents, whiteColor.cgColor.components)
@@ -30,5 +30,15 @@ class UIColorExtensionTest: XCTestCase {
         // test fail
         let failColor = UIColor.colorFrom(hex: "#F")
         XCTAssertEqual(failColor, UIColor.gray)
+    }
+    
+    func testColorFromHexRgb() {
+        let whiteCGCloroComponents: Array<CGFloat> = [1.0, 1.0, 1.0, 1.0],
+        whiteColor = UIColor.colorFrom(rgb: 0xFFFFFF)
+        XCTAssertEqual(whiteCGCloroComponents, whiteColor.cgColor.components)
+        
+        let whiteWithAlphaCGCloroComponents: Array<CGFloat> = [1.0, 1.0, 1.0, 0.5],
+        whiteColorWithAlpha = UIColor.colorFrom(rgb: 0xFFFFFF, alpha: 0.5)
+        XCTAssertEqual(whiteWithAlphaCGCloroComponents, whiteColorWithAlpha.cgColor.components)
     }
 }