+++ Source/UIColor+extension.swift
@@ -23,12 +23,15 @@ public extension UIColor {
         var rgbValue: UInt32 = 0
         Scanner(string: cString).scanHexInt32(&rgbValue)
         
-        return UIColor(
-            red: CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0,
-            green: CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0,
-            blue: CGFloat(rgbValue & 0x0000FF) / 255.0,
-            alpha: alpha
-        )
+        return colorFrom(rgb: rgbValue, alpha: alpha)
+    }
+    
+    static func colorFrom(rgb: UInt32, alpha: CGFloat = 1.0) -> UIColor {
+        let red = CGFloat((rgb & 0xFF0000) >> 16) / 0xff,
+            green = CGFloat((rgb & 0xFF00) >> 8) / 0xff,
+            blue = CGFloat(rgb & 0xFF) / 0xff
+        
+        return UIColor(red: red, green: green, blue: blue, alpha: alpha)
     }
 
 }